# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.

servers:
  - url: http://localhost:3000
    description: Development server

tags:
  - name: Leads
    description: Lead management endpoints

paths:
  /api/v1/leads:
    get:
      tags:
        - Leads
      summary: List Leads
      description: Get a paginated list of Leads with optional filtering and sorting
      operationId: listLeads
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeadList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Leads
      summary: Create Lead
      description: Create a new Lead
      operationId: createLead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLeadRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lead'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v1/leads/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'

    get:
      tags:
        - Leads
      summary: Get Lead by ID
      operationId: getLead
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lead'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Leads
      summary: Update Lead (full)
      operationId: updateLead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLeadRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lead'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - Leads
      summary: Update Lead (partial)
      operationId: patchLead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchLeadRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lead'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Leads
      summary: Delete Lead (soft)
      operationId: deleteLead
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v1/leads/bulk:
    post:
      tags:
        - Leads
      summary: Bulk create Leads
      operationId: bulkCreateLeads
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CreateLeadRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Lead'
    put:
      tags:
        - Leads
      summary: Bulk update Leads (full, atomic)
      operationId: bulkUpdateLeads
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                allOf:
                  - type: object
                    required: [id]
                    properties:
                      id:
                        type: string
                  - $ref: '#/components/schemas/UpdateLeadRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeadBulkResult'
        '400':
          $ref: '#/components/responses/BadRequest'
    patch:
      tags:
        - Leads
      summary: Bulk partial-update Leads (atomic)
      operationId: bulkPatchLeads
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  required: [ids, patch]
                  properties:
                    ids:
                      type: array
                      items: { type: string }
                    patch:
                      type: object
                      additionalProperties: true
                - type: object
                  required: [items]
                  properties:
                    items:
                      type: array
                      items:
                        type: object
                        required: [id, patch]
                        properties:
                          id: { type: string }
                          patch: { type: object, additionalProperties: true }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeadBulkResult'
        '400':
          $ref: '#/components/responses/BadRequest'

  /api/v1/leads/upsert:
    post:
      tags:
        - Leads
      summary: Upsert Lead
      operationId: upsertLead
      parameters:
        - name: key_field
          in: query
          schema:
            type: string
        - name: key_value
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLeadRequest'
      responses:
        '200':
          description: Upserted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lead'

  /api/v1/leads/trash:
    get:
      tags:
        - Leads
      summary: List deleted Leads
      operationId: listDeletedLeads
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeadList'

  /api/v1/leads/{id}/restore:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    post:
      tags:
        - Leads
      summary: Restore deleted Lead
      operationId: restoreLead
      responses:
        '200':
          description: Restored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lead'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v1/leads/empty:
    delete:
      tags:
        - Leads
      summary: Empty trash for Leads
      operationId: emptyTrashLeads
      responses:
        '200':
          description: Trash emptied
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted_count:
                    type: integer

  /api/v1/leads/delete/bulk:
    post:
      tags:
        - Leads
      summary: Bulk soft-delete Leads by ids (atomic)
      operationId: bulkDeleteLeads
      requestBody:
        $ref: '#/components/requestBodies/BatchIds'
      responses:
        '200':
          description: Soft-deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  soft_deleted: { type: integer }
        '400':
          $ref: '#/components/responses/BadRequest'

  /api/v1/leads/restore/bulk:
    post:
      tags:
        - Leads
      summary: Bulk restore Leads by ids (atomic)
      operationId: bulkRestoreLeads
      requestBody:
        $ref: '#/components/requestBodies/BatchIds'
      responses:
        '200':
          description: Restored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeadBulkResult'
        '400':
          $ref: '#/components/responses/BadRequest'

  /api/v1/leads/restore/all:
    post:
      tags:
        - Leads
      summary: Restore all soft-deleted Leads
      operationId: restoreAllLeads
      responses:
        '200':
          description: Restored
          content:
            application/json:
              schema:
                type: object
                properties:
                  restored: { type: integer }

  /api/v1/leads/trash/bulk:
    delete:
      tags:
        - Leads
      summary: Bulk permanently-delete Leads from trash by ids (atomic)
      operationId: bulkPermanentDeleteLeads
      requestBody:
        $ref: '#/components/requestBodies/BatchIds'
      responses:
        '200':
          description: Permanently deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  permanently_deleted: { type: integer }
        '400':
          $ref: '#/components/responses/BadRequest'

components:
  schemas:
    Lead:
      type: object
      required:
        - id
        - company_id
        - lead_name
        - source
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        lead_name:
          type: string
          maxLength: 140
        organization_name:
          nullable: true
          type: string
          maxLength: 140
        phone:
          nullable: true
          type: string
          maxLength: 30
        whatsapp_no:
          nullable: true
          type: string
          maxLength: 30
        email:
          nullable: true
          type: string
          maxLength: 160
        source:
          type: string
        campaign_id:
          nullable: true
          type: string
          format: uuid
        status:
          type: string
        party_id:
          nullable: true
          type: string
          format: uuid
        converted_at:
          nullable: true
          type: string
          format: date-time
        notes:
          nullable: true
          type: string
          maxLength: 2000
        owner_user_id:
          nullable: true
          type: string
          format: uuid
        sales_team_id:
          nullable: true
          type: string
          format: uuid
        merged_into_lead_id:
          nullable: true
          type: string
          format: uuid
        merged_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object

    LeadBulkResult:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Lead'
        total: { type: integer }
        failed: { type: integer }
        errors:
          type: array
          items: { type: string }

    CreateLeadRequest:
      type: object
      required:
        - company_id
        - lead_name
        - source
        - status
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        lead_name:
          type: string
          maxLength: 140
        organization_name:
          nullable: true
          type: string
          maxLength: 140
        phone:
          nullable: true
          type: string
          maxLength: 30
        whatsapp_no:
          nullable: true
          type: string
          maxLength: 30
        email:
          nullable: true
          type: string
          maxLength: 160
        source:
          type: string
        campaign_id:
          nullable: true
          type: string
          format: uuid
        status:
          type: string
        party_id:
          nullable: true
          type: string
          format: uuid
        converted_at:
          nullable: true
          type: string
          format: date-time
        notes:
          nullable: true
          type: string
          maxLength: 2000
        owner_user_id:
          nullable: true
          type: string
          format: uuid
        sales_team_id:
          nullable: true
          type: string
          format: uuid
        merged_into_lead_id:
          nullable: true
          type: string
          format: uuid
        merged_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object

    UpdateLeadRequest:
      type: object
      required:
        - company_id
        - lead_name
        - source
        - status
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        lead_name:
          type: string
          maxLength: 140
        organization_name:
          nullable: true
          type: string
          maxLength: 140
        phone:
          nullable: true
          type: string
          maxLength: 30
        whatsapp_no:
          nullable: true
          type: string
          maxLength: 30
        email:
          nullable: true
          type: string
          maxLength: 160
        source:
          type: string
        campaign_id:
          nullable: true
          type: string
          format: uuid
        status:
          type: string
        party_id:
          nullable: true
          type: string
          format: uuid
        converted_at:
          nullable: true
          type: string
          format: date-time
        notes:
          nullable: true
          type: string
          maxLength: 2000
        owner_user_id:
          nullable: true
          type: string
          format: uuid
        sales_team_id:
          nullable: true
          type: string
          format: uuid
        merged_into_lead_id:
          nullable: true
          type: string
          format: uuid
        merged_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object

    PatchLeadRequest:
      type: object
      properties:
        company_id:
          type: string
          format: uuid
        lead_name:
          type: string
          maxLength: 140
        organization_name:
          nullable: true
          type: string
          maxLength: 140
        phone:
          nullable: true
          type: string
          maxLength: 30
        whatsapp_no:
          nullable: true
          type: string
          maxLength: 30
        email:
          nullable: true
          type: string
          maxLength: 160
        source:
          type: string
        campaign_id:
          nullable: true
          type: string
          format: uuid
        status:
          type: string
        party_id:
          nullable: true
          type: string
          format: uuid
        converted_at:
          nullable: true
          type: string
          format: date-time
        notes:
          nullable: true
          type: string
          maxLength: 2000
        owner_user_id:
          nullable: true
          type: string
          format: uuid
        sales_team_id:
          nullable: true
          type: string
          format: uuid
        merged_into_lead_id:
          nullable: true
          type: string
          format: uuid
        merged_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object

    LeadList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Lead'
        meta:
          $ref: '#/components/schemas/PaginationMeta'

    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5

    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true

  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid

    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1

    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20

    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string

    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  requestBodies:
    BatchIds:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [ids]
            properties:
              ids:
                type: array
                items: { type: string }

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

security:
  - bearerAuth: []